hello_django/requirements.txt
asgiref==3.5.2
Django==3.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
whitenoise==5.3.0
gunicorn==20.1.0
dj-database-url==0.5.0
django-cors-headers==3.13.0
django-on-heroku==1.1.2